Refactored switch assignments function to new file
